[biosilico][crypto][review]Some move of files for view.
