--- biosilico-crypto/src/main/java/gabywald/crypto/chemical/data/package.html
@@ -1,8 +1,0 @@
-<HTML>
-<head><title>package 'gabywald.global.data.chemical'</title></head>
-<BODY>
-General chemical structures definitions.  
-<br>
-
-</BODY>
-</HTML>
