+++ biosilico-crypto/src/main/java/gabywald/crypto/data/chemical/Atom.java
@@ -1,4 +1,4 @@
-package gabywald.crypto.chemical.data;
+package gabywald.crypto.data.chemical;
 
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
