+++ biosilico-graphicalview/src/main/java/gabywald/crypto/launcher/CryptoLauncher.java
